Remove unnecessary methods
