+++ server/app.js
@@ -3,13 +3,7 @@ require('newrelic');
 const express = require('express');
 const routes = require('./../routes');
 const path = require('path');
-const bodyParser = require('body-parser');
-
 const app = express();
-
-// const jsonParser = bodyParser.json();
-// Middleware setup
-// app.use(jsonParser);
 
 app.set('port', process.env.PORT || 3003); //prod vs test environment switch
